MIGRATION: Add unique for column name on Drone, DroneCategory and Pilot Models.
