+++ drones/models/DroneCategory.py
@@ -1,0 +1,11 @@
+from django.db import models
+
+
+class DroneCategory(models.Model):
+    name = models.CharField(max_length=250, unique=True)
+
+    class Meta:
+        ordering = ('name', )
+
+    def __str__(self):
+        return self.name
