+++ drones/models/Pilot.py
@@ -1,0 +1,29 @@
+from django.db import models
+
+
+class Pilot(models.Model):
+
+    MALE = 'M'
+    FEMALE = 'F'
+    GENDER_CHOICES = (
+        (MALE, 'Male'),
+        (FEMALE, 'Female')
+    )
+    name = models.CharField(
+        max_length=150,
+        blank=False,
+        unique=True
+    )
+    gender = models.CharField(
+        max_length=2,
+        choices=GENDER_CHOICES,
+        default=MALE
+    )
+    races_count = models.IntegerField()
+    inserted_timestamp = models.DateTimeField(auto_now_add=True)
+
+    class Meta:
+        ordering = ('name', )
+
+    def __str__(self):
+        return self.name
