+++ drones/models/__init__.py
@@ -1,0 +1,4 @@
+from .Competition import Competition
+from .Drone import Drone
+from .DroneCategory import DroneCategory
+from .Pilot import Pilot
